Analysis of choice blindeness test.
We have some problems with singularity in the mixed logit model. 
Apparently, this has to do with the random intercept and random effects across items.  It seems to me that this is unrelated to other covariates because keeping just the random intercept for items results in singularity. I probably messed up the data simulation somewhere, which would not be an issue with real data. However, this can be something more substantial. I cannot spot the problem.