creating results
Results: perception of the behavior and person
One-off and repeated green behavior have an effect on evaluation of the person but not on perception of morality of immoral behavior.  The effect of one-off behavior seems to be larger (!!!) but I doubt this will be a significant difference. Maybe an interesting thing to test in another experiment.